[entropy_src/dv] Scoreboarding of REPCNT & REPCNTS
This commit fixes many subtle in scoreboard accuracies regarding the
behavior of the REPCNT and REPCNTS tests.

Moreover in order to bring the entropy_src_rng test to 100% passing
these changes assume the fixes proposed in issue #14078 and implemented
in PR #14083.

Additionally this PR addresses a scoreboarding issue wherein HT alerts
will not fire if the DUT main SM has been forced into an error state
(notably due to a local escalation, which is the only SM error that
the scoreboard can predict).

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>