refactor(kmac): Change `done` to `mubi4_t`
_Related Issue: https://github.com/lowRISC/opentitan/issues/13855_

`clear_i` triggers `rst_storage` in `keccak_round`. `rst_storage`
signal integrity is checked in the PR
https://github.com/lowRISC/opentitan/pull/14103

However, `clear_i` may be a victim of FI attacks. Then the consequences
are same to the FI attacks on `rst_storage`. Whenever the hashing
operation is completed and early `clear_i` signal is issued prior to SW
read, the result digest is 0. (It will be great if SW checks two shares
of the digests are not all zero)

This commit converts the `sha3_done` to `mubi4_t` and following logics
to check the signal to `prim_mubi_pkg::MuBi4True` strictly.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>