[entropy_src/rtl] Suppress watermarks after disable
Prior to this commit the HT watermarks are set to update even after
the IP is disabled.  This makes the watermark behavior different
from the alert and failure counter, because all of the test
failure pulses are masked once the IP is turned off.

This state of affairs means that the dataset used for watermarking
is effectively different then the data set used in calculating
failures and alerts, which creates big challenges for
verification.

By masking health_test_done_pulse with a copy of es enable,
this stops the watermarks from updating after disable and
makes the watermarks agree with the failure counters.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>