[bazel] use CRT for toolchain configuration (again)
PR #14081 tried to instroduce the CRT Bazel repo to manage toolchains
(over Bazel Embedded). However, it was quickly rolled back (in #14283)
after merging as it was incompatible with airgapped builds due to a
missing SHA256. This has since been resolved, and this change is once
again being instroduced.

Credit to @cfrantz for this update.

Signed-off-by: Timothy Trippel <ttrippel@google.com>