[entropy_src/rtl] Adjustment to main SM for FW_OV mode.
This commit makes it possible to jump straight from IDLE to the FWInsertMsg
state, sending a SHA3 start pulse in the process.  This is for cases when
the SHA3_Start signal in sent immediately after `module_enable` is inserted.
This change is neccessary as the module_enable signal passes through a 2-clock
synchronizer in the fanout module, and thus may be delayed to catch up with the
SHA3 start signal.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>