[otbn,rtl] Remove locking timing difference for secure wipe
When `otbn_controller` enters the *Locked* state during an initial
(i.e., post-reset) secure wipe, it immediately sets the `locking_o`
output. When `otbn_controller` enters the *Locked* state during a
post-execution secure wipe, however, it waits for the secure wipe to
complete before setting the `locking_o` output (see #14269).

This commit changes the `locking_o` signal for the initial secure-wipe
to wait for the secure wipe to complete and thereby aligns it with the
post-execution secure wipe.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>