feat(rdc): POK defined as a reset
This commit defines `vcaon_pok` as a reset (missed from auto recognition
by tool). POK is released right after Power is provided. So, the reset
scenario for the POK is to release with constraint that POR_N is still
in asserted.

As POK is introduced, POR_N has constraint with POK. There's no chance
POK and POR_N both are released.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>