[entropy_src/dv] Refine scoreboard handling of data
The SHA conditioner does not always get reset during a disable event.
When the scoreboard gets detected a disable event it needs to determine
which data will be purged as raw data, living in a FIFO vs. data
that has been absorbed in the SHA conditioning state.  In order to
simplify this these two classes of data are now stored in separate FIFOs.

This commit also adds some further timing adjustments to properly predict
a number of corner cases in dealing with processing of enable or disable
pulses close to SHA data inputs or SHA3 start commands.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>