[sw, otbn] Acknowledge Done interrupt in ECDSA P-256 test
Previously, we weren't clearing the state of the Done interrupt after
handling it. As a result, a false interrupt could get signaled
immediately after re-enabling it during the execution of the next
OTBN program which led Ibex to read out DMEM while OTBN was still busy
causing a fatal alert and test failures.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>