[entropy_src/rtl] Corner cases in handling delays
This commit handles a number of corner cases that can occur due to the
non-zero processing time of the SHA engine. Though all of these were
identified as by long delays in verifying the cs_aes_halt_req/ack
handshake, the following issues could all come up due to rapid
poorly-timed programming sequences.

The general issues are that:
1. The SHA engine is unresponsive to disable events while processing,
and
2. The bypass FIFO can stall for a number of reasons (BOOT mode
finished, mismatch between HT window size and seed size, etc).

The following corner-case delay scenarios are handled in this commit.

**Missed Disable Pulses**

The `main_sm` does not see disable pulses while the SHA block is
processing. (NOTE: For the purposes of this discussion AES-halt
handshakes are considered part of the SHA process time).  This
is by design as the SHA engine needs to complete it's sequence once
started. This can cause the module to remain in Continuous mode after
enabled rather than go back to Startup mode as required.

The correction here is to monitor for disable events that occur during
SHA processing and hold the `main_sm` disabled until processing has
completed. This forces the SM to return to Idle before proceeding.

If such stalls occur in the `main_sm` the rest of the core stalls as
well thanks to backpressure on the FIFOs

**Lingering SHA Outputs**

If SHA processing were instantaneous, a disable event would clear the
`esfinal` FIFO, and this FIFO would remain clear after re-enablement.
However if a disable event occurs during SHA processing, this ongoing
seed conditioning will continue, a possibly end up in the output FIFO.

For consistency and verifiability, these core now monitors for these
"stale" seeds, snd flushes them at the input of the `esfinal` FIFO.

**Configuration Change During SHA processing**

The core can toggle between bypass and non-bypass mode while disabled.
Since the core does not respond to disable pulses immediately, the
SHA3 block may also have inputs stalled waiting to be added to a
following seed.

Verification assumes that any 64-bits gathered in conditioning mode
will be instantly loaded into the conditioner, and will become part
of any following seed.  However if the conditioner stalls for processing
and the core is reconfigured the seed could be re-rerouted to the
bypass FIFO.

This mostly presents a big challenge for verification, However
maintainin gconsistency in this data-flow modele will help prevent
potential security issues.  This scenario merits a few changes:

1. The Pre-conditioning FIFO is now reserved for only conditioning data.
This lets the verification environment know that once any 64-bit word
has been submitted to the SHA3 engine, that word can considered already
aborbed regardless of delays.  Bypass data is now picked up directly at
the PostHT FIFO

2. Because many registers control the selection of bypass mode, the
signal `es_bypass_mode` is now a register which is only latched at
enable time.  This prevents misrouting of RNG data dues to transients
in configuration

**Non-Blocking Bypass FIFO**

The Bypass FIFO can block the rest of the pipeline when in bypass mode
for its own unique reasons.  If the bypass health-test window is larger
than a seed (or just a non-integer multiple of a seed), the Bypass FIFO
can become full before a full HT window.  Since bypass mode only
generates one seed per enable, these FIFO can also become full once
this single seed in generated.  For these reasons the Bypass FIFO is
left non-blocking.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>