[test] Enable passing arguments to the functest harness
Move the required args to the end and name them the "test commands."
Pass the test commands via an environment variable, and place them at
the end of the invocation.

This makes the test commands unable to be overridden by test_arg
arguments, and simultaneously, it allows the user to specify test_arg
arguments on the bazel command line to modify arguments to opentitantool
and friends. Those test_arg arguments can override or supplement the
test harness's arguments provided by the bazel rule.

Signed-off-by: Alexander Williams <awill@google.com>