feat(spid): Increase TPM RDFIFO depth to 16
This commit increases the TPM Read FIFO size to 64 Bytes (16 x 4B).

The previous impl of TPM Read FIFO relies on the SW latency. The SW
latency should not be long so that SW is able to write the Read data
continuously. In order to satisfy the assumption, SW shall turn off
interrupts.

The requirement above could not comply with SW architecture. So, this
commit increases the size for SW to write full 64 B at once.

Latter design change will follow. The design change is for HW to send
the TPM Read data only when full payload size is ready in the FIFO.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>