[otbn] Fix DMEM/IMEM rdata bus enabled assertions
These assertions tried to capture all conditions when a core would not
access DMEM or IMEM to assert that DMEM or IMEM would then be accessible
via the bus.  However, the conditions were not sufficiently specific,
and when DMEM or IMEM were not accessible while the conditions held, the
assertion would fail.  One option would be to make the conditions more
specific, but that would approach the definition of the enable signal,
which is not the point of an assertion.

Instead, an assertion should check if an implementation behaves as
specified.  OTBN's specification states that DMEM/IMEM can be read via
the bus when OTBN is *Idle*.

This commit changes those assertions to match the specification, thereby
closing #14602.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>